+++ src/lib/Components/ArtworkGrids/Artwork.tsx
@@ -20,7 +20,7 @@ const Badges = styled.View`
 `
 const Badge = styled.View`
   border-radius: 2px;
-  padding: 3px 5px 1px 5px;
+  padding: 3px 5px 1px 6px;
   background-color: white;
   margin-left: 5px;
 `
@@ -70,19 +70,20 @@ class Artwork extends React.Component<Props, any> {
   badges() {
     const { is_acquireable, is_biddable } = this.props.artwork
 
+    // TODO: Replace with proper size "0".
     return (
       this.hasBadges() && (
         <Badges>
           {is_acquireable && (
             <Badge>
-              <Sans weight="medium" size="1">
+              <Sans fontSize={8} lineHeight={8} style={{ paddingTop: 1 }} weight="medium" size="1">
                 BUY NOW
               </Sans>
             </Badge>
           )}
           {is_biddable && (
             <Badge>
-              <Sans weight="medium" size="1">
+              <Sans fontSize={8} lineHeight={8} style={{ paddingTop: 1 }} weight="medium" size="1">
                 BID
               </Sans>
             </Badge>